switch endpoint default to protocol generator
