+++ smithy-kotlin-codegen/src/main/kotlin/software/amazon/smithy/kotlin/codegen/rendering/protocol/HttpBindingProtocolGenerator.kt
@@ -19,8 +19,6 @@ import software.amazon.smithy.model.Model
 import software.amazon.smithy.model.knowledge.HttpBinding
 import software.amazon.smithy.model.shapes.*
 import software.amazon.smithy.model.traits.*
-import software.amazon.smithy.rulesengine.language.EndpointRuleSet
-import software.amazon.smithy.rulesengine.traits.EndpointTestCase
 import java.util.logging.Logger
 
 /**
@@ -953,36 +951,6 @@ abstract class HttpBindingProtocolGenerator : ProtocolGenerator {
         } else {
             // test if the request/response bindings have any members bound to the HTTP payload (body)
             bindings.any { it.location == HttpBinding.Location.PAYLOAD || it.location == HttpBinding.Location.DOCUMENT }
-        }
-    }
-
-    override fun generateEndpointProvider(ctx: ProtocolGenerator.GenerationContext, rules: EndpointRuleSet) {
-        val paramsSymbol = EndpointParametersGenerator.getSymbol(ctx.settings)
-        val providerSymbol = EndpointProviderGenerator.getSymbol(ctx.settings)
-        val defaultProviderSymbol = DefaultEndpointProviderGenerator.getSymbol(ctx.settings)
-
-        ctx.delegator.useFileWriter(paramsSymbol) {
-            EndpointParametersGenerator(it, rules).render()
-        }
-        ctx.delegator.useFileWriter(providerSymbol) {
-            EndpointProviderGenerator(it, paramsSymbol).render()
-        }
-        ctx.delegator.useFileWriter(defaultProviderSymbol) {
-            DefaultEndpointProviderGenerator(it, rules, providerSymbol, paramsSymbol).render()
-        }
-    }
-
-    override fun generateEndpointProviderTests(
-        ctx: ProtocolGenerator.GenerationContext,
-        tests: List<EndpointTestCase>,
-        rules: EndpointRuleSet,
-    ) {
-        val paramsSymbol = EndpointParametersGenerator.getSymbol(ctx.settings)
-        val defaultProviderSymbol = DefaultEndpointProviderGenerator.getSymbol(ctx.settings)
-        val testSymbol = DefaultEndpointProviderTestGenerator.getSymbol(ctx.settings)
-
-        ctx.delegator.useTestFileWriter("${testSymbol.name}.kt", testSymbol.namespace) {
-            DefaultEndpointProviderTestGenerator(it, rules, tests, defaultProviderSymbol, paramsSymbol).render()
         }
     }
 }
